Added logic to grant rights to the service account running the backup process also. Fix some framework thinning logic when bundling packages. Updated several exceptions to be custom.  Changed instance details to be InstanceCreationDetails to make a little more sense.

+++ Naos.Deployment.Core/PackageManager.cs
@@ -169,12 +169,33 @@ namespace Naos.Deployment.Core
                     var targetPath = Path.Combine(bundleStagePath, packageName);
                     ZipFile.ExtractToDirectory(packageFilePath, targetPath);
                     
-                    // thin out older frameworks so there is a single copy of the assembly (like if we have net45, net40, net35 - only keep net45...).
+                    // thin out older frameworks so there is a single copy of the assembly (like if we have net45, net40, net35, windows8, etc. - only keep net45...).
                     var libPath = Path.Combine(targetPath, "lib");
-                    var unnecessaryFrameworks = Directory.GetDirectories(libPath).OrderByDescending(_ => _).Skip(1).ToList();
-                    foreach (var unnecessaryFramework in unnecessaryFrameworks)
+                    var frameworkDirectories = Directory.Exists(libPath)
+                                                   ? Directory.GetDirectories(libPath)
+                                                   : new string[0];
+
+                    if (frameworkDirectories.Any())
                     {
-                        Directory.Delete(unnecessaryFramework, true);
+                        var frameworkFolderToKeep = frameworkDirectories.Length == 1
+                                                        ? frameworkDirectories.Single()
+                                                        : frameworkDirectories.Where(
+                                                            directoryPath =>
+                                                                {
+                                                                    var directoryName = Path.GetFileName(directoryPath);
+                                                                    var includeInWhere = directoryName != null
+                                                                                         && directoryName.StartsWith(
+                                                                                             "net",
+                                                                                             StringComparison.InvariantCultureIgnoreCase);
+                                                                    return includeInWhere;
+                                                                }).OrderByDescending(_ => _).FirstOrDefault();
+
+                        var unnecessaryFrameworks =
+                            frameworkDirectories.Except(new[] { frameworkFolderToKeep }).ToList();
+                        foreach (var unnecessaryFramework in unnecessaryFrameworks)
+                        {
+                            Directory.Delete(unnecessaryFramework, true);
+                        }
                     }
                 }
